Separate tests for cp and ucp
Rename commands again
