+++ crates/nu-command/src/filesystem/mod.rs
@@ -1,7 +1,6 @@
 mod cd;
 mod cd_query;
 mod cp;
-mod cp_old;
 mod glob;
 mod ls;
 mod mkdir;
@@ -11,6 +10,7 @@ mod rm;
 mod save;
 mod start;
 mod touch;
+mod ucp;
 mod util;
 mod watch;
 
@@ -18,7 +18,6 @@ pub use self::open::Open;
 pub use cd::Cd;
 pub use cd_query::query;
 pub use cp::Cp;
-pub use cp_old::CpOld;
 pub use glob::Glob;
 pub use ls::Ls;
 pub use mkdir::Mkdir;
@@ -27,4 +26,5 @@ pub use rm::Rm;
 pub use save::Save;
 pub use start::Start;
 pub use touch::Touch;
+pub use ucp::UCp;
 pub use watch::Watch;
