+++ crates/nu-command/tests/commands/mod.rs
@@ -100,6 +100,7 @@ mod to_text;
 mod touch;
 mod transpose;
 mod try_;
+mod ucp;
 mod uniq;
 mod uniq_by;
 mod update;
